requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
google-generativeai